2025/static-to-postgresql/post-db/sevenBPlus/requirements.txt
alembic==1.15.2
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
Flask==3.1.2
Flask-FlatPages==0.8.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Frozen-Flask==1.0.2
greenlet==3.2.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.10
Markdown==3.7
MarkupSafe==3.0.2
more-itertools==10.5.0
packaging==24.2
pip-review==1.3.0
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg-pool==3.2.6
pycparser==2.22
python-dotenv==1.1.0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
six==1.16.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3